Add Eval command
The Eval command in this implements a version of the quasi-REPL of
vim-fireplace.

I chose to rename EvalCode to EvalForm in the process, because I needed
a naming scheme that supports Eval as the "base" case, and EvalCode
sounds too much like something taking user input. EvalForm on the other
hand feels more natural for the functionality it provides; send form
under cursor to be evaluated.