Add prompting Doc command
For those cases where you want to look up the docs for a symbol that is
not currently in use.

The rename of Doc to DocCursor is to reflect the fact that Doc before
this commit was really a way to query the docstring of symbol under
cursor. Also the use of Doc for the prompting version of docstring
lookup matches the use of Eval for the prompting version of evaluating
code.