Use the new PREPL in 1.10
To make the results returned from the REPL parsable this commit switches
to a dependency on Clojure 1.10, where the PREPL (Program REPL) was
introduced.

This commit also switches away from explicitly using a PrintStream in the
repl-log namespace; instead a clojure.java.io/writer is used. This feels
more Clojur-y.

`tap>` will be used in conjunction with a custom tap listener to
implement communication with the REPL which should be hidden from the
user. That is a future commit.