Use `try` plus `finally` for the cp command
This makes it much more explicit that the channel needs to be closed
when the rest of the code is done executing. In reality this is still
just a proof of concept on how to use the internal socket repl to get a
result back to the plugin, but better a good example. ;-)