WIP: Add omnicompletion support
I debated two different approaches:

1. Inject a namespace which loads needed dependencies and uses
`clojure.repl/apropos` as a fallback
2. Use a custom classloader with the classpath of the repl project and
any dependencies

Ultimately I decided against option 2 because it has a lot more upkeep
with regards to dynamic evaluation in the repl; e.g. user evals a new
defn, that won't be picked up (I think?). The other point is that I know
too little about how classloaders work to feel comfortable using a
custom one.

Omnicomplete function requires the plugin to be able to return data to
the client. This was not possible with the async implementation of
`run-command`, so I renamed that to `run-command-async` and added
`run-command` as a sync alternative.

NOTE: This requires https://github.com/clojure-vim/neovim-client/pull/4